Emb_Testing/CI_Pipelines/arduino-due-pytest-hil-example-main/requirements.txt
pytest
xunitparser
pyserial